fix: garmin pipeline
