+++ garmin_pipeline.py
@@ -1,10 +1,5 @@
 """Process Fitness tracker data files"""
 
-import contextlib
-import os
-import tempfile
-import shutil
-from traceback import format_exc
 import zipfile
 
 import garmin.Garmin_Read_Sleep as garmin_read_sleep
@@ -17,14 +12,26 @@ import garmin.standard_respiratory_rate as garmin_standardize_respiratory_rate
 import garmin.standard_sleep_stages as garmin_standardize_sleep_stages
 import garmin.standard_stress as garmin_standardize_stress
 import garmin.metadata as garmin_metadata
+
+
+import argparse
+import os
+import tempfile
+import shutil
+import contextlib
+import time
+from traceback import format_exc
+import sys
+
 import azure.storage.filedatalake as azurelake
 import config
 import utils.dependency as deps
-from utils.file_map_processor import FileMapProcessor
-import utils.logwatch as logging
 import csv
-import time
+import utils.logwatch as logging
+from utils.file_map_processor import FileMapProcessor
 from utils.time_estimator import TimeEstimator
+from functools import partial
+from multiprocessing.pool import ThreadPool
 
 """
 # Usage Instructions:
@@ -57,41 +64,26 @@ for file in "$FitnessTracker_Path"/FitnessTracker-*/Garmin/Activity/*.fit \
 done
 """
 
+overall_time_estimator = TimeEstimator(1)  # default to 1 for now
 
-def pipeline(study_id: str):  # sourcery skip: low-code-quality
-    """Process fitness tracker data files for a study
-    Args:
-        study_id (str): the study id
-    """
 
-    if study_id is None or not study_id:
-        raise ValueError("study_id is required")
+def worker(
+    workflow_file_dependencies,
+    file_processor,
+    processed_data_output_folder,
+    manifest,
+    file_paths: list,
+    worker_id: int,
+):  # sourcery skip: low-code-quality
+    """This function handles the work done by the worker threads,
+    and contains core operations: downloading, processing, and uploading files."""
 
-    input_folder = f"{study_id}/pooled-data/FitnessTracker"
-    processed_data_output_folder = f"{study_id}/pooled-data/FitnessTracker-processed"
-    manifest_folder = f"{study_id}/pooled-data/FitnessTracker-manifest"
-    dependency_folder = f"{study_id}/dependency/FitnessTracker"
-
-    pipeline_workflow_log_folder = f"{study_id}/logs/FitnessTracker"
-    ignore_file = f"{study_id}/ignore/fitnessTracker.ignore"
-    manual_input_folder = f"{study_id}/pooled-data/FitnessTracker-manual"
-    red_cap_export_file = (
-        f"{study_id}/pooled-data/REDCap/AIREADiPilot-2024Sep13_EnviroPhysSensorInfo.csv"
+    logger = logging.Logwatch(
+        "fitness_tracker",
+        print=True,
+        thread_id=worker_id,
+        overall_time_estimator=overall_time_estimator,
     )
-    participant_filter_list_file = f"{study_id}/dependency/PatientID/AllParticipantIDs07-01-2023through07-31-2024.csv"
-
-    # input_folder = f"{study_id}/Stanford-Test/Pilot-Garmin/FitnessTracker-Pool"
-    # processed_data_output_folder = (
-    #     f"{study_id}/Stanford-Test/Pilot-Garmin/FitnessTracker-processed"
-    # )
-    # manifest_folder = f"{study_id}/Stanford-Test/Pilot-Garmin/FitnessTracker-manifest"
-    # dependency_folder = (
-    #     f"{study_id}/Stanford-Test/Pilot-Garmin/FitnessTracker-dependency"
-    # )
-
-    logger = logging.Logwatch("fitness_tracker", print=True)
-
-    # dev_allowed_list = ["1025", "7060", "4233", "1081", "4033", "4077", "7352"]
 
     # Get the list of blobs in the input folder
     file_system_client = azurelake.FileSystemClient.from_connection_string(
@@ -99,122 +91,8 @@ def pipeline(study_id: str):  # sourcery skip: low-code-quality
         file_system_name="stage-1-container",
     )
 
-    with contextlib.suppress(Exception):
-        file_system_client.delete_directory(processed_data_output_folder)
-
-    with contextlib.suppress(Exception):
-        file_system_client.delete_file(f"{dependency_folder}/file_map.json")
-
-    file_paths = []
-    participant_filter_list = []
-
-    # Create a temporary folder on the local machine
-    meta_temp_folder_path = tempfile.mkdtemp(prefix="garmin_pipeline_meta_")
-
-    # Get the participant filter list file
-    with contextlib.suppress(Exception):
-        file_client = file_system_client.get_file_client(
-            file_path=participant_filter_list_file
-        )
-
-        temp_participant_filter_list_file = os.path.join(
-            meta_temp_folder_path, "filter_file.csv"
-        )
-
-        with open(file=temp_participant_filter_list_file, mode="wb") as f:
-            f.write(file_client.download_file().readall())
-
-        with open(file=temp_participant_filter_list_file, mode="r") as f:
-            reader = csv.reader(f)
-            for row in reader:
-                participant_filter_list.append(row[0])
-
-        # remove the first row
-        participant_filter_list.pop(0)
-
-    paths = file_system_client.get_paths(path=input_folder)
-
-    logger.debug(f"Getting file paths in {input_folder}")
-
-    file_processor = FileMapProcessor(dependency_folder, ignore_file)
-
-    for path in paths:
-        t = str(path.name)
-
-        file_name = t.split("/")[-1]
-
-        # Check if the file name is in the format FIT-patientID.zip
-        if not file_name.endswith(".zip"):
-            logger.debug(f"Skipping {file_name} because it is not a .zip file")
-            continue
-
-        if len(file_name.split("-")) != 2:
-            logger.debug(
-                f"Skipping {file_name} because it does not have the expected format"
-            )
-            continue
-
-        cleaned_file_name = file_name.replace(".zip", "")
-
-        if file_processor.is_file_ignored(
-            cleaned_file_name, t
-        ) or file_processor.is_file_ignored(file_name, t):
-            logger.debug(f"Skipping {file_name}")
-            continue
-
-        patient_id = cleaned_file_name.split("-")[1]
-
-        # if str(patient_id) not in dev_allowed_list:
-        #     print(
-        #         f"dev-Participant ID {patient_id} not in the allowed list. Skipping {file_name}"
-        #     )
-        #     continue
-
-        if str(patient_id) not in participant_filter_list:
-            print(
-                f"Participant ID {patient_id} not in the allowed list. Skipping {file_name}"
-            )
-            continue
-
-        file_paths.append(
-            {
-                "file_path": t,
-                "status": "failed",
-                "processed": False,
-                "convert_error": True,
-                "output_uploaded": False,
-                "output_files": [],
-                "patient_folder_name": cleaned_file_name,
-                "patient_id": patient_id,
-            }
-        )
-
-    # Download the redcap export file
-    red_cap_export_file_path = os.path.join(meta_temp_folder_path, "redcap_export.tsv")
-
-    red_cap_export_file_client = file_system_client.get_file_client(
-        file_path=red_cap_export_file
-    )
-
-    with open(red_cap_export_file_path, "wb") as data:
-        red_cap_export_file_client.download_file().readinto(data)
-
-    # Create the output folder
-    file_system_client.create_directory(processed_data_output_folder)
-
-    workflow_file_dependencies = deps.WorkflowFileDependencies()
-
     total_files = len(file_paths)
-
-    logger.info(f"Found {total_files} items in {input_folder}")
-
     time_estimator = TimeEstimator(total_files)
-
-    manifest = garmin_metadata.GarminManifest(processed_data_output_folder)
-
-    manifest.read_redcap_file(red_cap_export_file_path)
-
-    # file_paths = file_paths[:5]
     for patient_folder in file_paths:
         patient_id = patient_folder["patient_id"]
 
@@ -867,6 +745,174 @@ def pipeline(study_id: str):  # sourcery skip: low-code-quality
 
             logger.time(time_estimator.step())
 
+
+def pipeline(study_id: str, workers: int = 4, args: list = None):
+    """The function contains the work done by
+    the main thread, which runs only once for each operation."""
+
+    if args is None:
+        args = []
+
+    global overall_time_estimator
+
+    # Process cirrus data files for a study. Args:study_id (str): the study id
+    if study_id is None or not study_id:
+        raise ValueError("study_id is required")
+
+    input_folder = f"{study_id}/pooled-data/FitnessTracker"
+    processed_data_output_folder = f"{study_id}/pooled-data/FitnessTracker-processed"
+    manifest_folder = f"{study_id}/pooled-data/FitnessTracker-manifest"
+    dependency_folder = f"{study_id}/dependency/FitnessTracker"
+
+    pipeline_workflow_log_folder = f"{study_id}/logs/FitnessTracker"
+    ignore_file = f"{study_id}/ignore/fitnessTracker.ignore"
+    manual_input_folder = f"{study_id}/pooled-data/FitnessTracker-manual"
+    red_cap_export_file = f"{study_id}/pooled-data/REDCap/AIREADiPilot-2024Sep13_EnviroPhysSensorInfo.csv"
+    participant_filter_list_file = f"{study_id}/dependency/PatientID/AllParticipantIDs07-01-2023through07-31-2024.csv"
+    logger = logging.Logwatch("fitness_tracker", print=True)
+
+    # dev_allowed_list = ["1025", "7060", "4233", "1081", "4033", "4077", "7352"]
+
+    # Get the list of blobs in the input folder
+    file_system_client = azurelake.FileSystemClient.from_connection_string(
+        config.AZURE_STORAGE_CONNECTION_STRING,
+        file_system_name="stage-1-container",
+    )
+
+    with contextlib.suppress(Exception):
+        file_system_client.delete_directory(processed_data_output_folder)
+
+    with contextlib.suppress(Exception):
+        file_system_client.delete_file(f"{dependency_folder}/file_map.json")
+
+    file_paths = []
+    participant_filter_list = []
+
+    # Create a temporary folder on the local machine
+    meta_temp_folder_path = tempfile.mkdtemp(prefix="garmin_pipeline_meta_")
+
+    # Get the participant filter list file
+    with contextlib.suppress(Exception):
+        file_client = file_system_client.get_file_client(
+            file_path=participant_filter_list_file
+        )
+
+        temp_participant_filter_list_file = os.path.join(
+            meta_temp_folder_path, "filter_file.csv"
+        )
+
+        with open(file=temp_participant_filter_list_file, mode="wb") as f:
+            f.write(file_client.download_file().readall())
+
+        with open(file=temp_participant_filter_list_file, mode="r") as f:
+            reader = csv.reader(f)
+            for row in reader:
+                participant_filter_list.append(row[0])
+
+        # remove the first row
+        participant_filter_list.pop(0)
+
+    paths = file_system_client.get_paths(path=input_folder)
+
+    logger.debug(f"Getting file paths in {input_folder}")
+
+    file_processor = FileMapProcessor(dependency_folder, ignore_file, args)
+
+    for path in paths:
+        t = str(path.name)
+
+        file_name = t.split("/")[-1]
+
+        # Check if the file name is in the format FIT-patientID.zip
+        if not file_name.endswith(".zip"):
+            logger.debug(f"Skipping {file_name} because it is not a .zip file")
+            continue
+
+        if len(file_name.split("-")) != 2:
+            logger.debug(
+                f"Skipping {file_name} because it does not have the expected format"
+            )
+            continue
+
+        cleaned_file_name = file_name.replace(".zip", "")
+
+        if file_processor.is_file_ignored(
+            cleaned_file_name, t
+        ) or file_processor.is_file_ignored(file_name, t):
+            logger.debug(f"Skipping {file_name}")
+            continue
+
+        patient_id = cleaned_file_name.split("-")[1]
+
+        # if str(patient_id) not in dev_allowed_list:
+        #     print(
+        #         f"dev-Participant ID {patient_id} not in the allowed list. Skipping {file_name}"
+        #     )
+        #     continue
+
+        if str(patient_id) not in participant_filter_list:
+            print(
+                f"Participant ID {patient_id} not in the allowed list. Skipping {file_name}"
+            )
+            continue
+
+        file_paths.append(
+            {
+                "file_path": t,
+                "status": "failed",
+                "processed": False,
+                "convert_error": True,
+                "output_uploaded": False,
+                "output_files": [],
+                "patient_folder_name": cleaned_file_name,
+                "patient_id": patient_id,
+            }
+        )
+
+    total_files = len(file_paths)
+
+    logger.info(f"Found {total_files} items in {input_folder}")
+
+    # Create the output folder
+    file_system_client.create_directory(processed_data_output_folder)
+    workflow_file_dependencies = deps.WorkflowFileDependencies()
+
+    # Download the redcap export file
+    red_cap_export_file_path = os.path.join(meta_temp_folder_path, "redcap_export.tsv")
+
+    red_cap_export_file_client = file_system_client.get_file_client(
+        file_path=red_cap_export_file
+    )
+
+    with open(red_cap_export_file_path, "wb") as data:
+        red_cap_export_file_client.download_file().readinto(data)
+
+    total_files = len(file_paths)
+
+    manifest = garmin_metadata.GarminManifest(processed_data_output_folder)
+
+    manifest.read_redcap_file(red_cap_export_file_path)
+
+    overall_time_estimator = TimeEstimator(total_files)
+
+    # Guarantees that all paths are considered, even if the number of items is not evenly divisible by workers.
+    chunk_size = (len(file_paths) + workers - 1) // workers
+    # Comprehension that fills out and pass to worker func final 2 args: chunks and worker_id
+    chunks = [file_paths[i : i + chunk_size] for i in range(0, total_files, chunk_size)]
+    args = [(chunk, index + 1) for index, chunk in enumerate(chunks)]
+    pipe = partial(
+        worker,
+        workflow_file_dependencies,
+        file_processor,
+        processed_data_output_folder,
+        manifest
+    )
+
+    # Thread pool created
+    pool = ThreadPool(workers)
+    # Distributes the pipe function across the threads in the pool
+    pool.starmap(pipe, args)
+
     file_processor.delete_out_of_date_output_files()
     file_processor.remove_seen_flag_from_map()
 
@@ -1025,4 +1071,19 @@ def pipeline(study_id: str):  # sourcery skip: low-code-quality
 
 
 if __name__ == "__main__":
-    pipeline("AI-READI")
+    sys_args = sys.argv
+
+    workers = 1
+
+    parser = argparse.ArgumentParser(description="Process garmin data files")
+    parser.add_argument(
+        "--workers", type=int, default=workers, help="Number of workers to use"
+    )
+    args = parser.parse_args()
+
+    workers = args.workers
+
+    print(f"Using {workers} workers to process garmin data files")
+
+    pipeline("AI-READI", workers, sys_args)
+
